+++ src/components/header/breadCrumbs.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'gatsby';
 
 type BreadCrumbsProps = {
   trail: Array<BreadCrumb>;
@@ -12,9 +13,9 @@ export interface BreadCrumb {
 const BreadCrumbs = ({ trail }: BreadCrumbsProps) => (
   <div className="w-full bg-lightBlue-light pl-4 py-1 ">
     {trail.map((breadCrumb, index: number) => (
-      <a href={breadCrumb.url} key={breadCrumb.name + index}>
+      <Link to={breadCrumb.url} key={breadCrumb.name + index}>
         <span className="text-blue text-xs">{`${breadCrumb.name} ${index < trail.length - 1 ? ' / ' : ''}`}</span>
-      </a>
+      </Link>
     ))}
   </div>
 );
