+++ src/templates/activityTemplate/index.tsx
@@ -14,7 +14,7 @@ interface ActivityPageTemplateProps {
 }
 
 const ActivityPageTemplate = ({ pageContext }: PageProps<object, ActivityPageTemplateProps>) => (
-  <Layout>
+  <Layout showBreadCrumbs>
     <div className="relative overflow-hidden h-86 mb-8">
       <div className="bg-gradient-to-t from-blue w-full h-full absolute opacity-75"></div>
       <img src={mockHero} className="w-full max-h-6/8 "></img>
